Add instruction to restrict encoding to four-byte characters
This adds the ability for users to only encode four-byte characters in a
given string. The motivation for adding this feature came from the fact
that MySQL's `utf8` character set doesn't include four-byte characters,
so it would be useful to be able to encode just those characters before
trying to persist them in a MySQL database.

https://dev.mysql.com/doc/refman/5.7/en/charset-unicode-utf8.html